semiautonomous-agents/a2a-protocol-dojo/agents/pyproject.toml
[project]
name = "a2a-dojo-agents"
version = "0.1.0"
description = "A2A Protocol Dojo - Demo Agents"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "uvicorn>=0.30.0",
    "starlette>=0.38.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
